✨ core: implement add command
